Update src/swiss_arrows/core.clj
Added a variation of the diamond wand which mostly behaves as the -> macro. I've used underscores as my _optional_ position indicator, which superficially seems non-breaking, and I like the aesthetics more, with '=>' being '->' with '_' :).

With the insertion point being optional, the => macro can be quite versatile, for example:

    (=> "foo" keyword {:bar _} {:baz _} [:fizz _ :buzz])

This is more a discussion point rather than a serious pull request, though I will not rule out the chance that you'll seriously consider it!